write iguana config to temp file and start benchmark

+++ benchmark-service/src/main/java/org/dicegroup/basilisk/benchmarkService/domain/iguana/task/TaskConfiguration.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Setter
 public class TaskConfiguration {
 
-    private Long timeLimit;
+    private Integer timeLimit;
 
     private QueryHandler queryHandler;
